+++ contracts/interfaces/hibachiStore.sol
@@ -2,5 +2,10 @@ pragma solidity ^0.6.2;
 pragma experimental ABIEncoderV2;
 
 interface HibachiStoreInterface {
-    function buyHibachi(address token, uint amt, address to, bool lottery) external;
+    function buyHibachi(
+        address token,
+        uint256 amt,
+        address to,
+        bool lottery
+    ) external;
 }
